energy_model: Memoize estimate_from_trace
The code in this module is _slow_. I had half-planned to optimize it,
but actually the pragmatic choice for now is just to use memoization
to avoid calling the slow code very much.

With this memoization I have estimate_from_trace taking 20s on a 78Mb
trace on a i5-2520M (quite a slow processor). I also tried memoizing
guess_idle_states but didn't find any further imrovement.

It may be possible to further speed this up by recognising the
special case that active_time == 1.0 or 0.0. Let's wait and see if
this turns out to be too slow.