analysis/latency: Fix last entry of latency_df
Similar to the overutilized dataframe, the last row of latency_df
will have a 't_delta' column value of 'NaN' because there is no
subsequent row to compute the delta from. This value is now the delta
between the timestamp of the last latency_df row and the end of the
trace.