utils/wa_results_collector: Fix opacity of legend colours
The trick in plot_comparisons which sets opacity on bars to represent p-value
has an issue in that the opacity of the first (lowest) bar sets the opacity of
the colours in the legend, which makes it very hard to read.

I don't think there's a nice way around this, so I've added a dummy set of bars
at the bottom of the plot, you can't see them because their value is 0, but
their opacity is 1, so the legend colours are fixed. Sorry, this is one of the
worst hacks I've ever committed, and it creates extra empty space on the plot
which I can't figure out how to get rid of.