tools/wa_results_collector: move stats generation into a dedicated function
This move the generation of statistics for a given metric into a dedicated
function describe(). The additional convenience method report() is added
to wrap boxplot and statistics generation in a single call.

The original do_boxplots() method has been renamed to be just boxplot to
better match API concepts already exposed by PANDAs.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>