tools/wltests: build: add modules_install option to build script
Some platforms need to be able to install modules on a target. In order
to do that, it is helpful to get the build system to produce module.deps
etc. files. We can ask the kernel makefile to install to a directory of
our choice by ensuring that we have INSTALL_MOD_PATH in set to something
in our environment. If you don't have that set, it defaults to '/' which
is normally what you do if installing modules to your local machine.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>