utils/wa_results_collector: allow user to select the sorting criteria
This allow the user to boxplot and report experiments ordered on a
specified statistic ('mean' by default) as well as with ascending or
descending order.

The ordering is kept in sync across a report() call so that boxplot
ordering alwasy matches tabular data.

If a non standard percentile is specified, an additional columns is generated
in the reported statistict.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>