tests: load_tracking: add PELT tests for tasks load tracking
Add a couple of PELT tests that check util_avg and load_avg of a task. For this
purpose we exploit the PELT simulator available in BART and compare the results
of simulation with the actual signal extracted from the trace.

In particular, the two tests check that the extracted signal's min, max and mean
values in the stable range correspond to the values of the simulated signal. The
second test looks at the behaviour of the signal and checks that it corresponds
to the one of the simulated signal.