tests/eas/generic: Add plot_expected_util
When running generics.py and trying to troubleshoot task_placement
failures, it can sometime be difficult to understand what was the
expected placement.

This commit adds a `plot_expected_util` method that will plot the
per-CPU utilization that was expected. It is called by default during
the test post-processing, and the plot will be saved as a .png file
in the test results output directory.