tests/eas/generic: Lower ramp tests duration
Each test in generic.py has a time duration set to 2s. However,
the "time_s" entry for rt-app ramps is the duration of each step,
and not the duration of the whole ramp. As such, each ramp test
lasts 26 seconds instead of 2.

This commit lowers the duration of each step to 500ms. Quoting @derkling:
 "
 If we consider a standard 32ms half-life PELT, a 70% task with
 16ms period can take up to 250ms to enter the stable region.
 "
This means the step duration can't be lowered too much, hence why
500ms seems a good enough choice.