test_series: fix the reboot procedure
An android device is considered as booted when the sys.boot_completed
field is set to 1 in getprop. The check that is done in the current
implementation of test_series has a logical bug, and a device is thus
considered as booted as soon as ADB pops up. This commit fixes the buggy
logic hence making the boot procedure more robust.