GenericFTrace: Add warning for accessing wrong attribute
We used to have a bug where when you had a known event whose 'name'
attribute != its 'unique_word', and you specified it explicitly in the
'events' param, we had two attributes - one for the 'name' and one for
the 'unique_word', and it was undefined which would be populated.
Now we just have the attribute from the 'name'. If there is any code
out there that was relying on this bug (i.e. accessing the
'unique_word' atribute), this will tell them what they need to do to
fix their code.

Co-author: Douglas Raillard <douglas.raillard@arm.com>