GenericFTrace: Fix overwriting of events in 'all' scope
Currently if you specify an event in the `events` by the `name` attribute of its
parser class, __add_events will add that name to the class_definitions dict. For
events in the 'thermal' and 'sched' scopes, that is then overwritten, but for
events in the dynamic_classes dict (where scope='all'), this is not the
case. That means that a specialised event parser class will be replaced with a
basic dynamic parser. Fix that by checking the 'name' field in known_events.