test_series: fix WA invocation with multiple devices
The ADB device name passed to test_series is currently used to flash and
reboot the board using ADB and Fastboot. However, it is not yet passed
to WA3 which then complains whenever several ADB devices are connected.
This commit fixes this behaviour by making the config fragment currently
used for the ACME more general and appending a "device_config" section
to it.