wlgen: add SchedEntity data structure
This is a data structure that can be used for creating hierarchies of sched
entities, namely tasks and taskgroups. It makes it easy to navigate through the
hierarchy, create workloads and run them in the specified Cgroup.

The hierarchy is represented as a tree where tasks can only be leaf nodes,
whereas taskgroup can have children, either tasks or taskgroups.