tests: load_tracking: add PELT tests for taskgroups utilization
This introduces a SchedEntity class which is a tree of tasks and taskgroups.
Tasks can only be leaves, whereas taskgroups can only be nodes.

The data structure makes it easy to come up with new hierarchies of tasks and
taskgroups and test that PELT utilization is correctly propagated through the
taskgroups.