tests/eas/generic: Fix matplotlib import
tests/eas/ was missing an __init__.py, which meant the backend set in
tests/__init__.py wouldn't be used.

This commit also removes the fill_between() call that is used to draw
the expected_placement figure, as it seems the desired behaviour is
not supported with the Agg backend ("Unknown property step").