tools/wltests: force exit in case of flashing issues
On some devices, flashing a new kernel can be a multi-stage operation
which can fail on some steps. In these cases, it does not make sense to
just skip the kernel and continue with the next one, which is likely
to fail for the same reasons.

This patch makes test_series completely failing in case an EIO error is
reported by one of the target specific flashing scripts. That script
is eventually required to warn the user and report the reason of the
failure in such a way that a further execution of test_series can
successfully flash new kernels.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>