wltests/wa_results_collector: Add composite energy for Monsoon results
Ideally, one would rewrite the monsoon devlib instrument to provide the
data in a form which allows the DerivedEnergyMeasurement class to give
summed energy & power values. However, the monsoon python driver does
not appear to be written in such a way that this could be done easily.

For the time being, lets make our own combosite measurement here if it
not supplied by Devlib and we have the known monsoon EM channels.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>