Promote pandas to a mandatory dependency.
Pandas is already widely used in WA and devlib for fast numerical table
manipulation in various instruments and output processors. Up to this
point, it has been an optional dependency.

Over the last couple of years, the package has become very wide spread,
and is available in repos for virtually every major distro. There is now
little reason not to require it as a mandatory dependency.