wa/config: specify the list of default trace events to collect
By default, WA's list of events to collect is the following:
        ['sched*', 'irq*', 'power*', 'thermal*']
As a result, the trace files are generally massive. This becomes an
issue when sched-evaluation-full-traced is in use as the result folder
can be > 25GB per kernel, mostly because of the trace files.

To avoid this, let's select by default a list of trace events we care
and drop all the others. This global parameter can be overridden in each
agenda if needed by setting 'trace_events' in the global section.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>