SchedEntity: add support to change the Taskgroup assigned to a Task
During the tests a Task may change from a Taskgroup to another.
It is then needed to update the parent of the Task and the
children of the old Taskgroup where the Task belonged and the children
of the new Taskgroup where the Task is included.
This patch adds as well a function that allows to search into the children of
a Taskgroup to find the object associate with a string name.

Signed-off-by: Elieva Pignat <Elieva.Pignat@arm.com>