load_tracking: add support to migrate a task between cgroups
For _PELTTaskGroupTest kind of tests, the support for allowing task to be
migrated from a cgroup to another during the execution has been added.
At midtime of the execution the test wakes up and allows the migration to
happen. The function _migrate_task can be overriden by Test to migrate
tasks. The time range where the utilization has been moved in the second
part of the execution in case a migration has occurred

Signed-off-by: Elieva Pignat <Elieva.Pignat@arm.com>