Using branch names in wltest series
Extend wltest with the capability of using, in addition to the SHA1,
also univocal branch names to checkout to a given kernel commit.
This allows to keep the tests updated with the HEAD commit of the given
branch name, without the need to update the series file every time the
HEAD of the branch changes.
If the branch name is not univocal, then return an error message is
returned and quit.