Check for kernel SHA1 to match also signed tags'
Since a signed tag has a SHA1 that differs from the one associated with
the commit to which it is referring, this was causing false negatives
when verifying the correspondence between the current kernel version of
the device and of the one to be tested.

This patch, when there is no correspondence between the current SHA1 of
the repository and the kernel of the device, verifies if there is a tag
pointing to the same SHA1 and performs the comparison with it.