rt-app: remove requirement for rooted device
We remove the requirement for rt-app to run on a rooted device as this
does not appear to be required. We make sure that we run the workload
from a directory where we have write permissions by first changing the
directory before executing the workload (in_directory argument of
invoke).

In the case that the device is rooted, we run rt-app as
root. Otherwise, we simply run the workload.