Docker: Add a Dockerfile for running WA
With this commit we add a Dockerfile which can be used to create a
Docker image for running WA and devlib. Users are able to use this
image in a Docker container to run workloads on devices, without
requiring WA on their host systems. Please be aware that udev rules
must still be correctly setup on the host.

NOTE: This Dockerfile will accept the Android SDK licenses
automatically.