wltests/wa_collector: add support for color bands to plot_comparisons
The plot_comparision is the most useful plot to get an overall view on
what are the relative changes in power/performances among different
kernels. However, to make more easy to have a fast check if a
certain kernel is ok or not, it can be useful to define a fixes set of
thresholds which allows to easily identify what's considered:
ok, suspicious or definitively not good.

Let's add the possibility to define these three thresholds and use them
to generate some color bands which allows to easily see where each
configuration stands.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>