CPU Migration test: fix parallelism issue
The file that contains the description of the tasks was written before
the test are launched and read back after the tests run to control the
that the result of the test are coherent.
The problem is when the test is run for several platforms in parallel
the file containing the json description may have been overwritten by
another platform test and may not contain the description of the test anymore.

To avoid the problem of sharing this file, the tasks descritpion is saved in
a variable in the test class and the file containing the json is no more read.

Signed-off-by: Elieva Pignat <Elieva.Pignat@arm.com>