rta.py: avoid file sharing
During the configuration a file is created with the content
of the workload to run. When tests on several platforms this file
is sharing between the different platforms resulting in concurrency
issues when the workloads are different for the different platforms.
To avoid the problem a unique temporary file is created during the
configuration of the workload. Such that the executions are no
more sharing the file. This file is deleted when it is no more needed.

Signed-off-by: Elieva Pignat <Elieva.Pignat@arm.com>