test_env: allow function profiling without events
This fixes:

   a73ed46f3899 env.py: Add debug_events key to target.config

where added a check to not generate an FtraceCollector if the user do
not specify some events. However, the functions profiler has not been
considered and we have also example notebooks where we want to be able
to profile kernel functions without necessarely specify trace events.

N.B. the FtraceCollector already provide the required logic to add at
least one event to properly enabled the function profiler. Look at that
devlib module for details.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>