tests/eas/generic.py: Add slack logging per task in _EnergyModelTest::_test_slack
The slack test case allows a task slack error percentage of 15%. In case
a test case passes it is impossible to know if it did so with 0% or
almost 15% error rate.

Log the slack percentage value per task to make even in case the task
stays under the slack error percentage to make it easier to post process
a test run.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>