utils/wa_results_collector: Add fallback sha1 acquisition method
To map a WA run with a kernel commit sha1, we currently read whatever
__meta/target_info.json reports for "kernel_release" (which should be
the equivalent of uname -r).

However, there are cases where this doesn't seem to be reported. For
instance, a HiKey960 built with hikey960_defconfig will not report the
commit sha1 in its version data, which will cause an exception in the
results collector code.

Since the sha1 is included in the results directory name, use that
as a backup.