utils/trace: make platform an optional parameter
Since:

   commit 074127b ("trace: Allow platform=None")

we allow to parse a Trace when platform data are not available.

However, the fix introduced by the previous patch was not changing the
order of the parameters to ensure backward compatibility.
This ended up in an (even more) ugly API where the first mandatory
parameter can be None.

Let's do the full cleanup and make platform as well as events two fully
optional parameters, thus reducing the trace parsing API to something
which in its based form can be:

   trace = Trace('/path/to/trace/file.dat')

Backward compatibility is not granted but this adds a sanity check to
detected usages of the old API and warn the user about the need to
update the call to the new convention.

Usages already in mainline of the Trace() API are instead fixed by this
patch by properly re-ordering or re-moving the required parameters.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>