energy_model: Add loader method to read simplified EM from target
If we cannot find an energy directory in the first domain & group of the
first online CPU, this target may be using the simplified EM which can be
loaded from many sources and provides a view of the loaded energy model
through a dedicated folder at /sys/devices/system/cpu/energy_model.

Provide a method to load from the new location and some utilities to try
to work out if one of them is present and load the correct one.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>