utils/trace: parse kernel version one time
When the latency analysis is performed on multiple tasks, for each one
we identify the kernel version and split out an annoying log message.

Since the kernel version is actually a trace property, let's parse it
from platform data (whenever available) before parsing the trace.
The latency analysis module is thus updated to use the pre-computed
kernel version.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>