tools/wltests: build: return immediately on failures
When a build target fails we do not properly forward the make exit code.
Let's to it so that, for example, we avoid to flash an image if a kernel
build should fails.

This is not expected to happen on properly tested and bisactable
SHA1 series, but this support is still valueable to have when for
example wltest is used to easly build/test cycles on development
kernels.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>