tests/cpuhp: Ensure tests fail when the target crashed
The CPU hotplug test is full of useless dependencies which silently talk
to the target (rt-app, freezing userspace, ftrace, ...). When the target
crashes during the hotplug stress, requests sent to the target because
of the above modules can hang indefinitely, or fail miserably. That is
an issue since, in the specific case of the hotplug test, we would like
to classify these crashes as test failures.

Remove all the dependencies by overwritting the runExperiments method of
LisaTest in order to avoid using an Executor. The test spawns a bash
script on the target that plugs CPUs in and out for a specific duration,
and check on the host side if the target is still alive after a while.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>