Added support for serialization and deserialization of the Energy Model.
This is a first step towards making new devices work out of the box (#425).
Take the Energy Model fetched from the target and transform it into a json file.
The ruamel.yaml package is used for the classes in energy_model.py.
Created static methods on the EnergyModel class, registered classes using decorators.