tests/preliminary: Skip SD_SHARE_CAP_STATES test when applicable
Recent EAS versions (with the simplified EM) do not rely on the
SD_SHARE_CAP_STATES flag to be set anywhere. As a result, the
preliminary test checking its presence fails, although the kernel
configuration is fine.

Fix this by making preliminary.py check the kernel version to skip
the test if it doesn't make sense.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>