Rework the wlgen API
The Workload class seems to have grown to mostly accomodate the RTA
subclass, and the perf workloads have grown unused.

This lead to a rewrite from scratch of the Workload class, with the
philosophy that it should be as simple as possible. In essence, this
class was used to execute a command on a target, and provide some
helpers around this command (cgroups, taskset...).

As such, this is all that the new Workload class does, on top of
saving the output of the command.

The RTA class has been re-arranged to fit this new incarnation of the
Workload class. It has not been entirely rewritten per-se, as a great
deal of code could be re-used. The philosophy for this rewrite was to
mostly eliminate the "magic dicts" (dicts used as parameters, instead
of explicit named parameters) that were laying around.

As such, the `RTA.conf()` step has been removed. The steps to run an
rt-app workload are now:

rta = RTA(...)
(or any alternate constructor such as rta = RTA.by_profile(...))
rta.run(...)