events: remove "print" event
The event matching infrastructure cannot cope with lines matching
multiple events, unless there are only 2 events matching and only one of
them is a fallback event. Unfortunately, TracingMarkWrite lines will
always contain a "print:" substring, which makes it ambiguous with the
Print event. Since no codebase is currently relying on the Print event,
it is safe to remove it.