python3: remove use of inspect.ismethod
In Python 3, class.method returns a function an not an unbound method as
it used to in Python 2. This makes inspect.ismethod(class.method) False,
so replace inspect.ismethod() with callable() builtin that will work in
both cases.

Also remove the use of dir() since it is meant for interactive use.
Replace with inspect.getmembers().

Improve the debugging experience by using a function instead of a lambda
to make it easier to find.