utils/env: Remove res_dir attribute
It doesn't make a lot of sense to force TestEnv instances to work
on a single result directory 'till their death. It would be better
to use a model where e.g. the workload code explicitely choses
a directory to write to, which is what this commit tries to
introduce.

The few caveats for that are:
- gem5: it needs a given directory for the host output. This needs to
  be passed through the TestEnv conf, or will default to a gem5
  directory under 'results'
- EM: it can take a directory, but just let it use its default. Energy
  data can be exported to a specific directory anyway.