lisa_shell: add a new function to generate buildroot rootfs
buildroot rootfs are a better choice to run lisa tests on as they are
quieter and provide a minimal userspace that allows testing kernel
functionality with minimum interference from userland. Also it is less
of a maintanence burden as it's unlikely to find unexpected things
configured without our knowledge that might intefere with some of the
tests or skew their results.

A recent example was found in systemd indirectly causing higher number
of kthreads to run on debian and arch-linux distro's which caused random
number of failure rates in cpu migration load tracking tests.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>