analysis: Fix __getattr__
Since __getattr__ is called by hasattr(), it may be called on an
un-initialized instance when being copied (copy module) or deserialized.
In order to avoid entering a loop when this code is probing for
__setstate__, we just bail out if we are requested a dunder name
(__*__). Such magic methods methods short-circuit the usual getattr()
lookup when it is done by builtins anyway, so the behavior should be
largely unchanged.