exekall: Add exekall
Add exekall and its integration in the lisa package.
When importing a module, exekall will look for a module named '<toplevel
package>.exekall_customize.adaptor' where the toplevel package is the
toplevel package of the module being imported. That module is
seggregated from the rest of the lisa package, so exekall is not
required to import any other part of lisa.