wa_results_collector.py: fix test_ids being None
When using the sched-evaluation-full.ipynb a weird exception about
unable to parse NaN values is thrown. This happens when running
example-jankbench.yaml agenda.

It turns out that since this agenda doesn't define test_ids, the code
tries to read the property anyway and sets it to None - which later
causes the exception when read from the ipython notebook.

Always check that the property is not empty before setting it - and
fallback to the workload name by default.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>