test_bundle: Add ResultBundle.display_and_exit
Add a display_and_exit() and exit method to be called in simple scripts
only executing one simple test.

Also add exekall option --callable-goal that is a pattern matching
callable names. This allows exporting our set of tests as a collection
of independent scripts, each executing one test and exiting through the
same helper. That is obviously much slower than running all tests under
exekall, but can be useful to check the structure of a test or to
execute them in another environment.