test_bundle: fix TestBundle.from_dir res_dir handling
Since TestBundle.from_dir function is mainly for interactive use, it
cannot make assumption on the location of the folder. That means that it
cannot take the decision to keep the root of the ArtifactPath or not.

Add an explicit parameter to allow not updating the res_dir for
serialization checks.