utils: Avoid registering the class in ruaml.yaml
Registering the class properly can only be done in a metaclass, which
would complicate the use of abc.ABC for example.

Also, not all classes are required to inherit from Serializable, which
means that many classes would not get the tag stability provided by
registering the class explicitely. Instead, the dotted name of the
classes are used.

Therefore, we avoid registering the class altogether, and will rely on
keeping stable class names when backward compatibilty is needed. That
can be done by keeping the old name available.

Existing yaml_tag class attributes are removed since they are no longer
exploited by ruamel.yaml