utils: Introduce multi-type support in MultiSrcConf
Allow a class tuple to be specified in the MultiSrcConf.STRUCTURE, like
the isinstance classinfo tuple.

Also use that in TargetConf to allow some keys to be present but None,
to make it possible to use a common target_conf.yml file depending on environment variable that may or may not be set.