utils/energy: Fix ACME for next
Switching to Python 3 broke the Popen-related code used for the ACME
support. Fix that by making sure to get a string as a result of
Popen.communicate().

While at it, introduce a rate limiting for the reset()/report() pattern
since iio-capture can have a funny behaviour if called/killed too
quickly.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>