lisa_shell: Check PYTHONPATH
Warn if PYTHONPATH is not empty, since it can interfere with venv. In
the worst case, it can lead to using a wrong version of e.g. bart that
would be installed somewhere else, leading to strange failures in LISA.

Also remove PYTHONPATH from the banner since it is not part of the
normal setup anymore.