bisector/exekall: drop xunit support
Now use exekall's ValueDB directly. This gives access to the full
ResultBundle in bisector, and removes some XML-related glue code.

Since the ValueDB is a tad big when the iteration count grows, only the
root values and exceptions are kept by default. This can be overriden
with "prune-db" option of exekall-LISA-test step class, if the user
wants to get a full DB when using -oexport-db.