lisa: utils: remove warning about LISA_HOME not set
Catching that at a such highlevel triggers the warning even when it is
completely irrelevant, such as generating the documentation. If the env
var is not set, the Python variable will be None and code making a use
of it as a string will fail anyway, so there is no risk of silent issue.