wltests: Introduce the WLTEST_REPO_TARGET option
Some targets (e.g. Pixel 3) require the utilization of Google's scripts
to build kernels and modules. Since this build system uses 'repo' and
manifests to manage the source trees and the toolchains directly, it
doesn't integrate easily in the existing wltest infrastructure which
kind of does that too.

In order to work around the issue, introduce a new option
(WLTEST_REPO_TARGET) which indicates that a platform uses the Google
build system. This option must be set to 'y' in the platform's
'definitions' file.

This option is basically used in test_series in order to checkout
commits of the repo manifest instead of kernel commits directly.
Whenever WLTEST_REPO_TARGET=y, the KERNEL_SRC env variable is supposed
to point at a repo folder, not at the kernel tree directly. The location
of the kernel tree within the repo is given in the platform code, using
the REPO_KERNEL_PATH variable.

While at it, clean-up  a few typos and unused arguments in test_series.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>