analysis/tasks: Add a stringify argument to df_task_states
The current df_task_states forces the user to use the

  df[df.curr_state == TaskState.TASK_*.char]

pattern since we store the state's character into the dataframe to be
readable. Analysis code shouldn't have to deal with this, and using
pure ints will make filtering a bit faster.

Change the existing df_task_states to use pure ints, and add a
parameter that lets us have the char representation of TaskState.