wlgen: Fix Ramp behavior
Make sure end_pct step is reached, by adding enough phases. If the
`delta_pct` is not an exact divider of `(end_pct-start_pct)`, we'll end
up with a final phase that has a duty cycle that does not match
`end_pct`. Force the last phase to have a duty cycle of `end_pct`.