bisector: disallow importing as __main__
Calling the script directly imports it as __main__, which breaks
references to classes expected to be hosted in a module called bisector.
That behavior is normal for a python application and handled by a shim
created by pip in a PATH location, so bisector needs to be called
through it.