lisa: Use PlatformInfo directly instead of TestEnv
Avoid using TestEnv when not necessary, since it forces a live
connection to a target. For example, getting the RTapp profile of a test
only requires accessing PlatformInfo, and we should be able to get it
without any live connection to a target.

Add sublevel 'kernel' to PlatformInfo, and a 'config' subkey.