git: Decode branch names and sha1
In find_shortest_symref() the subprocess command we use to get the list
of git branches returns a byte string. We then split this into byte
string 'name' and 'sha1' where this last will never match the SHA1
parameter passed in as a string.

Fix it be parsing the output of the subprocess command as normal
strings.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>