lisa/energy_model: Unify the way we load DT-EMs for both formats
We can write more generic code to handle both types of DT-based EM
without needing different code paths.

The two types differ in the amount of values each operating point
has in the model. The original format had a capacity,power pair
and the newer format has a capacity,frequency,power triple.

This version uses a grouper function to split the list
of EM strings into the required amount of sub-groups (determined
by dividing the number of values in cpu_capacity by the number
of reported operating points in the model) such that we can do
the following transformation:

[c0,p0,c1,p1,c2,p2] -> [[c0,p0],[c1,p1],[c2,p2]]
or
[c0,f0,p0,c1,f1,p1,c2,f2,p2] -> [[c0,f0,p0],[c1,f1,p1],[c2,f2,p2]]

We then use a lambda function in the iterator which is called
with each sub group list.

This function returns a new list containing only the first and
last item in the parameter. Since the first item in each list
is always cpu capacity and the last item in each list is always
power cost, we can pass this onto the iterator as we construct
the list of ActiveState objects for the EM.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>